+++ Assets/Scripts/Signals/InternalEvents/EnemySignals.cs
@@ -8,7 +8,7 @@ namespace Signals
     public class EnemySignals
     {
         public UnityAction<StickmanBodyPartEnums> onDie = delegate { };
-        public UnityAction<int, StickmanBodyPartEnums> onHitted = delegate { };
+        public UnityAction<float, StickmanBodyPartEnums> onHitted = delegate { };
         public UnityAction onShoot = delegate { };
         public UnityAction onReload = delegate { };
         public UnityAction onReloaded = delegate { };
